wip - refactor scripts
